Added many comments
Our accuracy is lower than it should be but I think it was so good before because the way the tokenizer had been trained wrong by mistake. We're getting 50% accuracy and while in reality that not bad I feel like we could change things to make it better. A good test would be to remove the sigmoid function, change the changing the size of the biLSTM layers and double check the changed tokenization and train test slpit.