Add files via upload
Can someone please look at the confusion matrix and try and fix it. I have no idea what is wrong. Also the third section with this warning "<ipython-input-3-05e137f6c983>:2: SettingWithCopyWarning: 
A value is trying to be set on a copy of a slice from a DataFrame.
Try using .loc[row_indexer,col_indexer] = value instead"

I've worked on this for hours and I'm not getting anywhere.